Add support for backend telemetry (tracing & metrics). Cherry picked from b792c7a4c7393acd51e9a7069635ca7f896d2f9d
This uses the opentelemetry SDK to provide support for traces and two
simple metrics (a conversations count and a messages count).

The telemetry entrypoint is built separately (unfortunately it does not
seem like there is a method for doing this in one build as sveltekit
overrides rollupOptions).

As written the metrics may be too high cardinality for some use cases
(as they are keyed per user), but these attributes can always be dropped
in a collector.